requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
autograd==1.3
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.6
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
decorator==4.4.2
flit_core==3.8.0
fonttools==4.39.2
frozenlist==1.3.3
future==0.18.3
google-auth==2.6.0
google-auth-oauthlib==0.4.4
grpcio==1.42.0
idna==3.4
imageio==2.9.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
kiwisolver==1.2.0
lazy_loader==0.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.2
networkx==3.0
numpy==1.22.3
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pip==23.0.1
pooch==1.4.0
protobuf>=3.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==23.0.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2023.2
PyWavelets==1.1.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.20.0
scipy==1.8
setuptools==65.6.3
six==1.16.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2020.7.24
torch==1.12.1
torchvision==0.13.1
tqdm==4.48.2
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wheel==0.38.4
win-inet-pton==1.1.0
wincertstore==0.2
yarl==1.8.1
zipp==3.11.0
